pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scam_detection"
version = "0.1.0"
description = "ML pipeline for scam detection"
authors = [{name = "Eugene Zolotarev"}]
requires-python = ">=3.9"
dependencies = [
    # NOTE: torch/torchvision are intentionally not pinned here.
    # They are installed in the Docker image as ROCm wheels.
    "lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "dvc>=3.0.0",
    "mlflow>=2.0.0",
    "transformers>=4.20.0",
    "datasets>=2.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0,<2.0.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "fire>=0.5.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
scam_detection = "scam_detection.commands:main"

[tool.hatch.build.targets.wheel]
packages = ["scam_detection"]

[dependency-groups]
dev = [
    "pre-commit>=3.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "mypy>=1.0.0",
]
